redesign & performances done

+++ SmartTask.Web/Services/Implementations/BaseService.cs
@@ -1,4 +1,4 @@
-﻿using System.Linq.Expressions;
+using System.Linq.Expressions;
 using SmartTask.Web.Infrastructure.Interfaces;
 using SmartTask.Web.Services.Interfaces;
 
@@ -13,41 +13,39 @@ namespace SmartTask.Web.Services.Implementations
             IGenericRepository<T> repository,
             IUnitOfWork unitOfWork)
         {
-            _repository = repository;
-            _unitOfWork = unitOfWork;
+            _repository = repository ?? throw new ArgumentNullException(nameof(repository));
+            _unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
         }
 
         public virtual async Task<IEnumerable<T>> GetAllAsync()
-        {
-            return await _repository.GetAllAsync();
-        }
+            => await _repository.GetAllAsync();
 
         public virtual async Task<T?> GetByIdAsync(int id)
-        {
-            return await _repository.GetByIdAsync(id);
-        }
+            => id <= 0 ? null : await _repository.GetByIdAsync(id);
 
         public virtual async Task<IEnumerable<T>> FindAsync(Expression<Func<T, bool>> predicate)
-        {
-            return await _repository.FindAsync(predicate);
-        }
+            => predicate == null ? Enumerable.Empty<T>() : await _repository.FindAsync(predicate);
 
         public virtual async Task AddAsync(T entity)
         {
+            ArgumentNullException.ThrowIfNull(entity);
             await _repository.AddAsync(entity);
             await _unitOfWork.SaveChangesAsync();
         }
 
         public virtual async Task UpdateAsync(T entity)
         {
+            ArgumentNullException.ThrowIfNull(entity);
             _repository.Update(entity);
             await _unitOfWork.SaveChangesAsync();
         }
 
         public virtual async Task DeleteAsync(int id)
         {
-            var entity = await _repository.GetByIdAsync(id);
+            if (id <= 0)
+                return;
 
+            var entity = await _repository.GetByIdAsync(id);
             if (entity != null)
             {
                 _repository.Delete(entity);
